bump protobuf, and partition process.ex to fix warnings
